move helping functions to contrib
